use 64-bit integer for power level values
While there are hardly any larger power level values than a
32-bit integer can hold in practice, the specification requires
values as large as 2^53-1, so 64-bit integer values are required,
if we want to conform to the specification.

+++ Mocktrix.Events.Tests/room/PowerLevelsEventTests.cs
@@ -130,7 +130,7 @@ namespace Mocktrix.Events.Tests
                 Content = new PowerLevelsEventContent
                 {
                     Ban = 50,
-                    Events = new SortedDictionary<string, int>()
+                    Events = new SortedDictionary<string, long>()
                     {
                         { "m.room.name", 100 },
                         { "m.room.power_levels", 100 }
@@ -138,13 +138,13 @@ namespace Mocktrix.Events.Tests
                     EventsDefault = 0,
                     Invite = 50,
                     Kick = 50,
-                    Notifications = new SortedDictionary<string, int>()
+                    Notifications = new SortedDictionary<string, long>()
                     {
                         { "room", 20 }
                     },
                     Redact = 50,
                     StateDefault = 50,
-                    Users = new SortedDictionary<string, int>()
+                    Users = new SortedDictionary<string, long>()
                     {
                         { "@example:localhost", 100 }
                     },
